+++ back/web-app/src/login/login.module.ts
@@ -3,7 +3,7 @@ import { LoginController } from './login.controller';
 import { LoginService } from './login.service';
 import { ConfigModule } from '@nestjs/config';
 import { FtStrategy } from './ft.strategy';
-import { UserModule } from 'src/users/user.module';
+import { UserModule } from 'src/users/user/user.module';
 import { SessionModule } from 'src/session/session.module';
 
 @Module({
