+++ back/web-app/src/signup/signup.controller.ts
@@ -1,5 +1,5 @@
 import { Body, Controller, Post } from '@nestjs/common';
-import { UserService } from 'src/users/user.service';
+import { UserService } from 'src/users/user/user.service';
 import { Builder } from 'builder-pattern';
 import { CreateMemberUserDto } from 'src/users/dto/create-member-user.dto';
 import { User } from 'src/users/entities/user.entity';
