+++ back/web-app/src/signup/signup.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { SignupController } from './signup.controller';
 import { SignupService } from './signup.service';
-import { UserModule } from 'src/users/user.module';
+import { UserModule } from 'src/users/user/user.module';
 
 @Module({
   imports: [UserModule], // Use to UserService
